Fixed player ID type and began using PacketBuffer instead of WebSocketBuffer.

+++ src/net/threesided/server/Player.java
@@ -8,22 +8,22 @@ public class Player extends Circle {
      * The type of player.
      */
     public enum Type {
-        TREE(0),
-        SNOWMAN(1);
+        TREE((byte) 0),
+        SNOWMAN((byte) 1);
 
-        private final int id;
+        private final byte id;
 
         /**
          * @param id The ID of the type.
          */
-        Type(final int id) {
+        Type(final byte id) {
             this.id = id;
         }
 
         /**
          * @return The Type's ID.
          */
-        public int getID() {
+        public byte getID() {
             return this.id;
         }
 
@@ -31,7 +31,7 @@ public class Player extends Circle {
          * @param id The ID of the type.
          * @return The Type based on the given ID.
          */
-        public static Type getByID(final int id) {
+        public static Type getByID(final byte id) {
             switch (id) {
                 case 0:
                     return TREE;
@@ -49,7 +49,7 @@ public class Player extends Circle {
     private static final int DEFAULT_RADIUS = 20;
     private static final int DEFAULT_MASS = 5;
 
-    private final byte id;
+    private final short id;
     private byte lives;
     private Type type;
     private String username;
@@ -60,7 +60,7 @@ public class Player extends Circle {
      * @param username The player's username.
      * @param lives The number of lives the player should initially have.
      */
-    public Player(final byte id, final Player.Type type, final String username, final byte lives) {
+    public Player(final short id, final Player.Type type, final String username, final byte lives) {
         super(Player.DEFAULT_RADIUS);
         this.id = id;
         this.type = type;
@@ -86,7 +86,7 @@ public class Player extends Circle {
     /**
      * @return The player's ID.
      */
-    public byte getID() {
+    public short getID() {
         return this.id;
     }
